rename utils.py to util.py
