+++ src/executer.py
@@ -5,7 +5,7 @@ import copy
 import numpy as np
 
 from .set import set_networks, set_device
-from .utils.utils import video_writer, load_video, adjust_imsize, concat_all, extend_rec, CreateRandMask, check_mask_position, find_bbox
+from .utils.util import video_writer, load_video, adjust_imsize, concat_all, extend_rec, CreateRandMask, check_mask_position, find_bbox
 from .utils.mask_design import create_mask, center_mask, edge_mask, create_boxline, write_boxline
 from .utils.auxiliary_layer import max_mask_size
 from .processer import GANonymizer
